+++ test/python/synthesis/aqc/test_aqc.py
@@ -41,7 +41,7 @@ class TestAqc(QiskitTestCase):
 
         seed = 12345
 
-        num_qubits = int(round(np.log2(ORIGINAL_CIRCUIT.shape[0])))
+        num_qubits = round(np.log2(ORIGINAL_CIRCUIT.shape[0]))
         cnots = make_cnot_network(
             num_qubits=num_qubits, network_layout="spin", connectivity_type="full", depth=0
         )
